+++ packages/superdough/helpers.mjs
@@ -1,4 +1,4 @@
-import { getAudioContext } from './webaudio.mjs';
+import { getAudioContext } from './superdough.mjs';
 
 export function gainNode(value) {
   const node = getAudioContext().createGain();
